V-10: 1. Fields for entering contacts are moved to a separate component; 2. Added the ability to remove the specialization input field.
