+++ src/screens/ContactsModalScreen/ContactsModalScreen.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { View, Modal, Text } from 'react-native';
-import { styles, inputFieldStyles } from './styles';
-import { InputField } from '../../components/InputFields/InputField/InputField';
+import { styles } from './styles';
 import { ButtonStrip } from '../../components/ButtonStrip/ButtonStrip';
+import { ContactsList } from '../../components/ContactsList/ContactsList';
 
 export const ContactsModalScreen = () => {
   return (
@@ -11,26 +11,10 @@ export const ContactsModalScreen = () => {
         <View style={styles.container}>
           <View style={styles.contactsContainer}>
             <Text style={styles.headerTitle}>Спасибо за решение помочь!</Text>
-            <InputField
-              title="Оставьте свой телефон"
-              placeholder="+7 (___) ___-__-__"
-              extStyle={inputFieldStyles}
-              keyboardType="phone-pad"
-            />
-            <InputField
-              title="или e-mail"
-              placeholder="Введите e-mail"
-              extStyle={inputFieldStyles}
-              keyboardType="email-address"
-            />
-            <InputField
-              title={`и укажите специализацию,\nв которой можете помочь`}
-              placeholder="Укажите специализацию"
-              extStyle={inputFieldStyles}
-            />
+            <ContactsList />
             <Text style={styles.fieldTitle}>Мы с вами свяжемся.</Text>
           </View>
-          <ButtonStrip actionText="ПЕРЕВЕСТИ" />
+          <ButtonStrip actionText="ОТПРАВИТЬ" />
         </View>
       </View>
     </Modal>
